+++ mcpython/client/rendering/blocks/TemporaryChestRenderer.py
@@ -22,5 +22,5 @@ class TemporaryChestRenderer(
 ):
     def add(self, position: typing.Tuple[int, int, int], block, face, batches):
         return shared.model_handler.add_raw_face_to_batch(
-            position, {}, None, batches, face
+            block, position, {}, None, batches, face
         )